DBZ-3889 Moving ContainerImageVersions to core for broader re-use and using it in two more places
