+++ debezium-testing/debezium-testing-testcontainers/pom.xml
@@ -49,6 +49,11 @@
       <groupId>org.testcontainers</groupId>
       <artifactId>mongodb</artifactId>
     </dependency>
+    <dependency>
+      <groupId>io.debezium</groupId>
+      <artifactId>debezium-core</artifactId>
+      <type>test-jar</type>
+    </dependency>
 
     <dependency>
       <groupId>org.easytesting</groupId>
